Change: move delete_user to dedicated users files
