+++ src/manage_users.h
@@ -77,4 +77,7 @@ create_user (const gchar *, const gchar *, const gchar *, const gchar *,
              int, const array_t *, array_t *, gchar **,
              array_t *, gchar **, gchar **, user_t *, int);
 
+int
+delete_user (const char *, const char *, int, const char *, const char *);
+
 #endif /* not _GVMD_MANAGE_USERS_H */
